Move argparse extensions to separate module
